refactor: normalize color used from css vars
