+++ console/packages/starwhale-ui/src/theme/styletron.ts
@@ -1,0 +1,6 @@
+import { createThemedStyled, createThemedWithStyle, createThemedUseStyletron } from 'baseui'
+import { SWThemeT } from './index'
+
+export const themedStyled = createThemedStyled<SWThemeT>()
+export const themedWithStyle = createThemedWithStyle<SWThemeT>()
+export const themedUseStyletron = createThemedUseStyletron<SWThemeT>()
