+++ console/src/components/Table/TableTyped.tsx
@@ -15,16 +15,15 @@ import {
     Types,
 } from '@starwhale/ui/base/data-table'
 import _ from 'lodash'
-// import useResizeObserver from '@/hooks/window/useResizeObserver'
 import CategoricalTagsColumn from '@starwhale/ui/base/data-table/column-categorical-tags'
 import { areEqual } from 'react-window'
 import type { RowT } from '@starwhale/ui/base/data-table/types'
 import { IStore } from '@starwhale/ui/base/data-table/store'
 import { useEffect } from 'react'
-import { useStyletron } from 'baseui'
 import { createUseStyles } from 'react-jss'
 import cn from 'classnames'
 import BusyPlaceholder from '../BusyLoaderWrapper/BusyPlaceholder'
+import { themedUseStyletron } from '@starwhale/ui/theme/styletron'
 
 const useStyles = createUseStyles({
     table: {
@@ -109,7 +108,6 @@ export function TableTyped({
             return {
                 ...column,
                 fillWidth: false,
-                // maxWidth: 200,
             }
         }
 
@@ -204,7 +202,7 @@ export function TableTyped({
         }
     }, [$columns])
 
-    const [, theme] = useStyletron()
+    const [, theme] = themedUseStyletron()
     const styles = useStyles({ theme })
 
     return (
@@ -245,7 +243,7 @@ export function TableTyped({
                         },
                         TableHeadCell: {
                             style: {
-                                backgroundColor: 'var(--color-brandTableHeaderBackground)',
+                                backgroundColor: theme.brandTableHeaderBackground,
                                 fontWeight: 'bold',
                                 borderBottomWidth: 0,
                                 fontSize: 14,
