+++ console/vite.config.ts
@@ -50,7 +50,7 @@ export default defineConfig({
     },
     resolve: { alias },
     plugins: [
-        // eslint(),
+        eslint(),
         react({
             exclude: /\.stories\.(t|j)sx?$/,
         }),
